+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/GanttDependency.cs
@@ -5,9 +5,9 @@ namespace Kendo.Mvc.Examples.Models
 {
     public partial class GanttDependency
     {
-        public long ID { get; set; }
-        public long PredecessorID { get; set; }
-        public long SuccessorID { get; set; }
+        public int ID { get; set; }
+        public int PredecessorID { get; set; }
+        public int SuccessorID { get; set; }
         public long Type { get; set; }
     }
 }
