+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/GanttResource.cs
@@ -5,7 +5,7 @@ namespace Kendo.Mvc.Examples.Models
 {
     public partial class GanttResource
     {
-        public long ID { get; set; }
+        public int ID { get; set; }
         public string Color { get; set; }
         public string Name { get; set; }
     }
