+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/OrderDetail.cs
@@ -5,10 +5,10 @@ namespace Kendo.Mvc.Examples.Models
 {
     public partial class OrderDetail
     {
-        public long OrderID { get; set; }
-        public long ProductID { get; set; }
+        public int OrderID { get; set; }
+        public int ProductID { get; set; }
         public double Discount { get; set; }
-        public long Quantity { get; set; }
+        public int Quantity { get; set; }
         public double UnitPrice { get; set; }
 
         public virtual Order Order { get; set; }
