+++ wrappers/mvc-6/demos/Kendo.Mvc.Examples/Models/SamplesEntities/Territory.cs
@@ -11,7 +11,7 @@ namespace Kendo.Mvc.Examples.Models
         }
 
         public string TerritoryID { get; set; }
-        public long RegionID { get; set; }
+        public int RegionID { get; set; }
         public string TerritoryDescription { get; set; }
 
         public virtual ICollection<EmployeeTerritory> EmployeeTerritories { get; set; }
